Fixed incorrect TLevel type to avoid warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@58903 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/data_loaders/genbank/dispatcher.hpp
@@ -64,7 +64,7 @@ public:
     typedef  CReader::TBlobId                   TBlobId;
     typedef  CReader::TChunkId                  TChunkId;
     typedef  CReader::TChunkIds                 TChunkIds;
-    typedef  int                                TLevel;
+    typedef  size_t                             TLevel;
     typedef  vector<CSeq_id_Handle>             TIds;
 
     CReadDispatcher(void);
